12 Sep 2021 update
Massive clean up + reliability checks.

+++ lib/SQL-comments.sql
@@ -1,7 +1,7 @@
 CREATE TABLE `comments` (
-  `comment_id` int(20) NOT NULL,
-  `user_id` int(11) NOT NULL,
-  `id` int(20) NOT NULL,
+  `comment_id` bigint(20) NOT NULL,
+  `user_id` bigint(20) NOT NULL,
+  `id` bigint(20) NOT NULL,
   `timestamp` datetime NOT NULL DEFAULT current_timestamp(),
   `message` text NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
@@ -13,4 +13,4 @@ ALTER TABLE `comments`
   ADD KEY `timestamp` (`timestamp`);
 
 ALTER TABLE `comments`
-  MODIFY `comment_id` int(20) NOT NULL AUTO_INCREMENT;
+  MODIFY `comment_id` bigint(20) NOT NULL AUTO_INCREMENT;
