+++ lib/SQL-forgot.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `password_reset` (
-  `user_id` int(11) NOT NULL,
+  `user_id` bigint(20) NOT NULL,
   `reset_hash` varchar(64) NOT NULL,
   `reset_time` datetime NOT NULL
 ) ENGINE=InnoDB DEFAULT CHARSET=utf8;
