+++ lib/SQL-reacts.sql
@@ -1,6 +1,6 @@
 CREATE TABLE `reactions` (
-  `id` int(11) NOT NULL,
-  `user_id` int(11) NOT NULL,
+  `id` bigint(20) NOT NULL,
+  `user_id` bigint(20) NOT NULL,
   `reaction` tinyint(1) NOT NULL DEFAULT '1'
 ) ENGINE=InnoDB DEFAULT CHARSET=latin1;
