+++ lib/SQL-users.sql
@@ -1,5 +1,5 @@
 CREATE TABLE `users` (
-  `user_id` int(11) NOT NULL,
+  `user_id` bigint(20) NOT NULL,
   `user_name` varchar(255) NOT NULL,
   `user_email` varchar(255) NOT NULL,
   `user_password` varchar(255) NOT NULL
@@ -11,4 +11,4 @@ ALTER TABLE `users`
   ADD KEY `user_name` (`user_name`);
 
 ALTER TABLE `users`
-  MODIFY `user_id` int(11) NOT NULL AUTO_INCREMENT;
+  MODIFY `user_id` bigint(20) NOT NULL AUTO_INCREMENT;
